+++ src/entities/Error.ts
@@ -1,3 +1,5 @@
+import { JsonrpcId } from 'src/entities';
+
 export class JsonrpcError extends Error {
   code: number;
 
@@ -7,11 +9,26 @@ export class JsonrpcError extends Error {
   }
 }
 
-export class WebsocketError extends Error {
-  code: number;
+export const INVALID_JSON_REQUEST = {
+  jsonrpc: '2.0',
+  id: null,
+  error: {
+    code: -32700,
+    message: 'invalid json format',
+  },
+};
 
-  constructor(message: string, code: number) {
-    super(message);
-    this.code = code;
-  }
-}
+export const customRpcError = (
+  message: string,
+  args?: {
+    id?: null | JsonrpcId;
+    code?: number;
+  },
+) => ({
+  jsonrpc: '2.0',
+  id: args?.id ?? null,
+  error: {
+    code: args?.code ?? -32700,
+    message,
+  },
+});
